# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vinmo API",
        "description": "Documentação da API do Vinmo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/notifications/sse-ticket": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Obtém um ticket para conexão Server-Sent Events (SSE)",
                "operationId": "4833c7f6d950b93462606585b7e9d134",
                "responses": {
                    "200": {
                        "description": "Ticket obtido com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Lista as notificações do usuário",
                "operationId": "6d56a555955089a73fdf3f7229e8ead8",
                "responses": {
                    "200": {
                        "description": "Notificações obtidas com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/{id}/read": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Marca uma notificação como lida",
                "operationId": "cde64b28769413444ae4ae3c3de1bff5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da notificação",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificação marcada como lida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/read-all": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Marca todas as notificações como lidas",
                "operationId": "a938b29776d5df701355702f21a24149",
                "responses": {
                    "200": {
                        "description": "Notificações atualizadas com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/{id}/dismiss": {
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Remove uma notificação",
                "operationId": "0d3519c4807504e5f2b257ec73721d43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da notificação",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificação removida com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/stream": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Stream de notificações via SSE",
                "operationId": "a0dd9a008b949544c5284d565026bf7d",
                "responses": {
                    "200": {
                        "description": "Stream estabelecida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}